Delete the endpoints module
